+++ rsl_rl_leggedloco/rsl_rl_leggedloco/modules/actor_critic_recurrent.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 import torch
 import torch.nn as nn
 
-from rsl_rl.modules.actor_critic import ActorCritic, get_activation
-from rsl_rl.utils import unpad_trajectories
+from rsl_rl_leggedloco.modules.actor_critic import ActorCritic, get_activation
+from rsl_rl_leggedloco.utils import unpad_trajectories
 
 
 class ActorCriticRecurrent(ActorCritic):
