+++ rsl_rl_leggedloco/setup.py
@@ -4,7 +4,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="rsl_rl",
+    name="rsl_rl_leggedloco",
     version="2.0.2",
     packages=find_packages(),
     author="ETH Zurich, NVIDIA CORPORATION",
